safari-534.53-branch/LayoutTests/http/tests/loading/preload-append-scan-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
slow-script.pl has MIME type application/javascript
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
preload-test.jpg has MIME type image/jpeg
main frame - didFinishLoadForFrame

This test needs to be run in DRT. Preload scanner should see the image resource.
<img src="resources/preload-test.jpg">
